Move WebSocketService out of Server in preparation
